refactor: create folder for source files
