--- .editorconfig
@@ -1,9 +1,0 @@
-# top-most EditorConfig file
-root = true
-
-[*]
-charset = utf-8
-insert_final_newline = true
-indent_style = tab
-indent_size = 4
-tab_width = 4
