+++ src/Magic.ts
