+++ src/tsconfig.json
@@ -1,0 +1,23 @@
+{
+  "compilerOptions": {
+	  "baseUrl": "..",
+	  "inlineSourceMap": true,
+	  "inlineSources": true,
+	  "module": "ESNext",
+	  "target": "ES6",
+	  "allowJs": true,
+	  "noImplicitAny": true,
+	  "moduleResolution": "node",
+	  "importHelpers": true,
+	  "isolatedModules": true,
+	  "lib": [
+      "DOM",
+      "ES5",
+      "ES6",
+      "ES7"
+    ]
+  },
+  "include": [
+    "**/*.ts"
+  ]
+}
