+++ src/version-bump.mjs
@@ -1,4 +1,4 @@
-import { readFileSync, writeFileSync } from "fs";
+import {readFileSync, writeFileSync} from "fs";
 
 const targetVersion = process.env.npm_package_version;
